Split capability events into focused daemon submodules
